Fix: fixed some bugs connected to refactor
